EATDer/requirements.txt
absl-py==2.0.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
filelock==3.9.0
fonttools==4.45.1
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
huggingface-hub==0.19.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lightning-utilities==0.9.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.3.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
timm==0.9.10
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.16.2+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
